Title: 사각형 채우기 3, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241117/사각형 채우기 3/rectangle-fill-3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
-int arr[1001] = {0};
-int sum[1001] = {0};
+long long arr[1001] = {0};
+long long sum[1001] = {0};
 int n;
 
 int main() {
